eatgo-customer-api/src/main/java/kr/co/papercraft/eatgo/application/RegionService.java
package kr.co.papercraft.eatgo.application;

import kr.co.papercraft.eatgo.domain.Model.Region;
import kr.co.papercraft.eatgo.domain.Repository.RegionRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class RegionService {

    @Autowired
    RegionRepository regionRepository;

    public RegionService(RegionRepository regionRepository) {
        this.regionRepository = regionRepository;
    }

    public List<Region> getRegions() {
        return regionRepository.findAll();
    }

}
